Generalized SignModel handling
