+++ src/main/java/ru/bclib/client/render/BaseSignBlockEntityRenderer.java
@@ -1,6 +1,5 @@
 package ru.bclib.client.render;
 
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import com.mojang.blaze3d.platform.NativeImage;
 import com.mojang.blaze3d.vertex.PoseStack;
@@ -9,16 +8,14 @@ import com.mojang.math.Vector3f;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityModelLayerRegistry;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
-import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.ModelLayers;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
-import net.minecraft.client.renderer.blockentity.SignRenderer;
+import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import net.minecraft.client.renderer.blockentity.SignRenderer.SignModel;
-import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.resources.model.Material;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
@@ -27,47 +24,36 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.SignBlock;
 import net.minecraft.world.level.block.StandingSignBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.WoodType;
 import net.minecraft.world.phys.Vec3;
 import ru.bclib.blockentities.BaseSignBlockEntity;
 import ru.bclib.blocks.BaseSignBlock;
+import ru.bclib.client.SignModelFactory;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
 
 public class BaseSignBlockEntityRenderer implements BlockEntityRenderer<BaseSignBlockEntity> {
 	private static final HashMap<Block, RenderType> LAYERS = Maps.newHashMap();
-	private static final Set<WoodType> TYPES = new HashSet<>();
 	private static final RenderType defaultLayer;
 	private final Font font;
-	private final Map<WoodType, SignModel> signModels;
-	private final SignModel defaultModel;
+	private final SignModelFactory signModelFactory;
+
 
 	private static final int OUTLINE_RENDER_DISTANCE = Mth.square(16);
 
-	public BaseSignBlockEntityRenderer(EntityRendererProvider.Context ctx, ModelLayerLocation signLayerLocation) {
+	public BaseSignBlockEntityRenderer(BlockEntityRendererProvider.Context ctx) {
 		super();
-
-		//build a list of all new sign models.
-		this.signModels = (Map) TYPES.stream().collect(ImmutableMap.toImmutableMap((signType) -> {
-			return signType;
-		}, (signType) -> {
-			return new SignRenderer.SignModel(ctx.bakeLayer(ModelLayers.createSignModelName(signType)));
-		}));
-
-		//set up a default model
-		defaultModel = new SignRenderer.SignModel(ctx.bakeLayer(ModelLayers.createSignModelName(WoodType.OAK)));
-
+		this.signModelFactory = new SignModelFactory(ctx.getModelSet());
 		this.font = ctx.getFont();
 	}
 
 	public void render(BaseSignBlockEntity signBlockEntity, float tickDelta, PoseStack matrixStack,
 			MultiBufferSource provider, int light, int overlay) {
 		BlockState state = signBlockEntity.getBlockState();
-		WoodType woodType = getSignType(state.getBlock());
-		SignRenderer.SignModel model = this.signModels.get(woodType);
+		SignModel model = signModelFactory.getSignModel(state);
 
 		matrixStack.pushPose();
 
@@ -135,6 +121,8 @@ public class BaseSignBlockEntityRenderer implements BlockEntityRenderer<BaseSign
 		matrixStack.popPose();
 	}
 
+
+
 	private static boolean isOutlineVisible(BaseSignBlockEntity signBlockEntity, int i) {
 		if (i == DyeColor.BLACK.getTextColor()) {
 			return true;
@@ -151,19 +139,10 @@ public class BaseSignBlockEntityRenderer implements BlockEntityRenderer<BaseSign
 		}
 	}
 
-	public static WoodType getSignType(Block block) {
-		WoodType signType2;
-		if (block instanceof SignBlock) {
-			signType2 = ((SignBlock) block).type();
-		} else {
-			signType2 = WoodType.OAK;
-		}
 
-		return signType2;
-	}
 
 	public static Material getModelTexture(Block block) {
-		return Sheets.getSignMaterial(getSignType(block));
+		return Sheets.getSignMaterial(SignModelFactory.getSignType(block));
 	}
 	
 	public static VertexConsumer getConsumer(MultiBufferSource provider, Block block) {
@@ -173,7 +152,7 @@ public class BaseSignBlockEntityRenderer implements BlockEntityRenderer<BaseSign
 	public static void registerCustomModel(Block block, EntityModelLayerRegistry.TexturedModelDataProvider provider) {
 		ResourceLocation blockId = Registry.BLOCK.getKey(block);
 		SignType type = new SignType(blockId.getPath());
-		TYPES.add(type);
+		SignModelFactory.TYPES.add(type);
 
 		EntityModelLayerRegistry.registerModelLayer(ModelLayers.createSignModelName(type), provider);
 	}
